Setup tests with v2 module
