+++ ridemyway/tests/tests_v1/test_create_ride.py
@@ -5,7 +5,7 @@
 import unittest
 import json
 
-from . import BaseTest
+from ridemyway.tests import BaseTest
 from . import (VALID_RIDE_DATASET,
                INVALID_DATE_DATASET,
                PAST_DATE_DATASET,
