+++ ridemyway/tests/tests_v1/test_ride_request.py
@@ -5,7 +5,7 @@
 import unittest
 import json
 
-from . import BaseTest
+from ridemyway.tests import BaseTest
 from . import VALID_RIDE_DATASET, VALID_RIDE_DATASET_1
