+++ ridemyway/tests/tests_v2/__init__.py
@@ -1,0 +1,21 @@
+"""
+    BASE testing module
+"""
+from ridemyway.tests import BaseTest
+
+
+class BaseTest(BaseTest):
+    """
+        Base class for testing
+    """
+    def setUp(self):
+        """
+            Set up tests
+        """
+        super().setUp()
+
+    def tearDown(self):
+        """
+            Teardown all test files and instances created
+        """
+        super().tearDown()
