+++ ridemyway/tests/tests_v2/test_auth.py
@@ -1,0 +1,1 @@
+from ridemyway.tests import BaseTest
